refactor: split assetsPlugin into separate export
- Move `vitePluginFullstack` and `serverHandlerPlugin` to `src/plugins/fullstack.ts`
- Create `AssetsPluginOptions` type for `assetsPlugin`
- Add `src/assets.ts` with `assetsPlugin` as default export
- Export `/assets` subpath for standalone usage

Closes #3
