+++ src/assets.ts
@@ -1,0 +1,1 @@
+export { assetsPlugin as default } from "./plugin.ts";
