+++ src/index.ts
@@ -1,1 +1,2 @@
-export { default, serverHandlerPlugin, assetsPlugin } from "./plugin.ts";
+export { default, serverHandlerPlugin } from "./plugins/fullstack.ts";
+export { assetsPlugin } from "./plugin.ts";
